Improve result ranking
Honestly, it's still not great. The best next step appears to be taking
the list of inventory items and using `fuzzywuzzy` directly for more
control over how the query is scored against each item.

In short, `sphobjinv` does enough for a proof-of-concept, but how its
`.suggest()` method works isn't quite search-engine-y enough.