Add better logging to show derivation of base URL
That "one-liner" assigning directly to bot.memory was clever, but it
didn't tell the bot owner whether the base URL comes from `link_base` in
the config or assumed from the `inventory` setting.

Now the logs do that, at the cost of a few extra LoC. They also say when
a trailing slash is added automatically.

Also changed `LOGGER.exception()` calls in `update_sphinx_objects()` to
`LOGGER.error()`; the tracebacks are not useful.